Add wireframe placeholder overlay, LOD building filter, and tile count overflow fix
Add SDF wireframe grid overlay on placeholder tiles for visual feedback
while tiles load. Skip building geometry decode at zoom levels below 13
where buildings are invisible, saving CPU and GPU resources. Fix integer
overflow in tile_visible.c when computing total tile count at high zoom
levels.

+++ client/src/tile_visible.c
@@ -16,7 +16,7 @@ int arpt_enumerate_visible_tiles(const arpt_camera *cam, int level,
        enumerate every tile.  This sidesteps all ray-casting edge cases
        (limb under-sampling, antimeridian, polar convergence) and
        guarantees no visible tile is ever missed. */
-    int total_tiles = n_cols * n_rows;
+    int64_t total_tiles = (int64_t)n_cols * n_rows;
     if (total_tiles <= max_count) {
         int count = 0;
         for (int y = 0; y < n_rows; y++)
